Move type converter to own namespace
