+++ src/IbanNet/Iban.cs
@@ -4,6 +4,7 @@ using System.ComponentModel;
 using System.Diagnostics;
 using System.Linq;
 using System.Text.RegularExpressions;
+using IbanNet.TypeConverters;
 
 namespace IbanNet
 {
